+++ test-packages/test-services/v4/test-service/TestEntity.js
@@ -338,6 +338,15 @@ var TestEntitySingleLink_1 = require('./TestEntitySingleLink');
     true
   );
   /**
+   * Static representation of the [[enumPropertyInt64]] property for query construction.
+   * Use to reference this property in query operations such as 'select' in the fluent request API.
+   */
+  TestEntity.ENUM_PROPERTY_INT_64 = _fieldBuilder.buildEdmTypeField(
+    'EnumPropertyInt64',
+    'Edm.Enum',
+    true
+  );
+  /**
    * Static representation of the [[enumPropertyWithOneMember]] property for query construction.
    * Use to reference this property in query operations such as 'select' in the fluent request API.
    */