+++ test-packages/test-services/v4/test-service/TestEntity.ts
@@ -494,6 +494,15 @@ export namespace TestEntity {
     true
   );
   /**
+   * Static representation of the [[enumPropertyInt64]] property for query construction.
+   * Use to reference this property in query operations such as 'select' in the fluent request API.
+   */
+  export const ENUM_PROPERTY_INT_64 = _fieldBuilder.buildEdmTypeField(
+    'EnumPropertyInt64',
+    'Edm.Enum',
+    true
+  );
+  /**
    * Static representation of the [[enumPropertyWithOneMember]] property for query construction.
    * Use to reference this property in query operations such as 'select' in the fluent request API.
    */